client/challenge_solver - move InfobloxClient to plugins
