+++ acmetk/client/__init__.py
@@ -1,10 +1,9 @@
 from .client import AcmeClient
-from .challenge_solver import InfobloxClient, DummySolver, ChallengeSolver
+from .challenge_solver import DummySolver, ChallengeSolver
 from .exceptions import CouldNotCompleteChallenge, AcmeClientException
 
 __all__ = [
     "AcmeClient",
-    "InfobloxClient",
     "DummySolver",
     "CouldNotCompleteChallenge",
     "ChallengeSolver",
